+++ CHANGELOG.md
@@ -7,6 +7,10 @@ Adheres to [Semantic Versioning](http://semver.org/).
 ## 6.3.0 (TBD)
 
 * geopackage-android version 6.3.0
+* LatLngBoundingBox coordinate constructor and directional methods
+* MapUtils build click and tolerance method enhancements (density, zoom, pixel bounds)
+* FeatureInfoBuilder feature styles fine filter support
+* FeatureOverlayQuery style pixel bounds support for querying visualized geometry styles and icons
 
 ## [6.2.1](https://github.com/ngageoint/geopackage-android-map/releases/tag/6.2.1) (03-11-2022)